+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Upgrade the `org.realityforge.arez` artifacts to version `0.174`.
 * Upgrade the `org.realityforge.arez.testng` artifacts to version `0.06`.
 * Upgrade the `org.realityforge.react4j` artifacts to version `0.154`.
+* Rework the way `EntityBrokerLock` so that multiple locks can be acquired by different connectors simultaneously. Only when all locks are released will the broker re-enable or resume.
 
 ### [v6.86](https://github.com/replicant4j/replicant/tree/v6.86) (2020-04-27) · [Full Changelog](https://github.com/replicant4j/replicant/compare/v6.85...v6.86)